Change behavior in utils.import_object().
Fix bug 972859.

Prior to this patch, utils.import_object(foo) and
utils.import_class(foo) would return the same thing if foo was the path
to a class.  This patch changes utils.import_object() to return an
instance of the class, instead, also allowing you to pass in arguments
to the constructor.

This patch also removes the odd behavior of import_object() to be able
to also import a module.  Instead, a new function import_module() has
been added.

Finally, the three import helpers have been moved to a new module,
importutils.

Change-Id: Ia2ea64c965692f8c3a29adec332dd93a83980070