Update exception from importutils.import_class().
Update the message included in the exception from
importutils.import_class() to include what the inner exception was.
This is quite helpful as the real error may not have been that the class
was not found.  It may have failed to import the module that contains
that class, and it's nice to know that and why that happened.

Change-Id: I085864d230b46adce8921e362f058f9421a5a674