Copy global requirements on update.py run.
Copy any entry from the global list of dependencies that exists in the
project's pip- or test-requires files. This way we can keep the versions
the same and ensure that projects only carry a subset of the total set
of packages without adding a bunch of extra depends to everyone.

Fixes bug 995607

Change-Id: I1215df190d8781117ebe511c291cd65b5a4d9248