Add impl_zmq to rpc.
Part of blueprint common-rpc.

This driver just went into nova.rpc.  Pull it in over here.  I haven't
been able to test this, but I can't get the tests to work in nova,
either.

There is a bit of a complication with impl_zmq worth mentioning.  It
requires a service to run (nova/bin/nova-rpc-zmq-receiver).  Only once
instance of this service should run on a given machine, so it shouldn't
be copied into each project that uses rpc.  Once openstack-common is an
installed library, it can be distributed with that, but until then ...
we'll just leave it in nova for now.

Change-Id: I2bd067be58d943d4f3a90e013b22282217484872