Avoid erroneous "Unsupported RPC Version" message if method is missing
The recent RPC versioning support introduced some odd behavior if a non-
existent method is called. Instead of reporting that the method does not
exist (as the code used to do, in the form of an AttributeError), it now
Reports "Unsupported RPC Version". This is highly confusing, as the RPC
versions *do* match and is just a result of the logic used to check
every advertised RPC proxy for the method.

This patch notes if any of the RPC versions matched the requested one,
and if so, raises an AttributeError if the method is never found,
instead of UnsupportedRpcVersion.

Also add unit tests for the cases where a method is missing, both when
the API version matches and not.

Change-Id: Icb7b13fa890000f1bd03382d5cce747e3311ef6a