More robust mailmap parsing.
The project has two different versions of this floating around. The one in
nova just splits on space and doesn't work with python-novaclient .mailmap
which has full names in some of its entries. The one in python-novaclient
tries splitting with an if, but it breaks in nova. This one should properly
get the mapping built properly.

Change-Id: I32974957494728e4fcc1667e301026dcd7fb8004