Don't switch to list_notifier unless necessary.
Some notification tests depend on a different, single notifier being
used.  So, don't switch to the list notifier just because we loaded
this class; only switch if an actual plugin needs it.

Also:  improved the test that makes sure all this notifier-switching
doesn't break the notifier that was present in the first place.

Change-Id: I170477bccb3ad9505fe6dbfbc214f77c09f6862f