Rework the version consumer functionality.
Moved the version routines into a class for better structure.
Added caching/memoization logic so that pkg_resources/git doesn't get
called more than is necessary.
Added periodic-based caching that will only cause the meta info refs to
be checked at most once a day, to reduce network load for developers. (The
information in the meta refs should really only change once a month, and
this will only affect local developer workstation version reporting, so we
should be fine here.
Added a deferred version string class and function to allow for consuming of
the version string but doing the calculation only when it is requested,
based on test runs in glance.

Change-Id: I6a436c8817f6904959727a76d876b807ec545ee9