Prevent weird double ~'s in local versions.
The version generation code didn't account for developers running locally.
So while checking for the entry in meta/openstack/release that matches
the branch name works for release builds, it left an empty string in the
milestonever section of the version name. Instead, if we just poke the
branch name in there, local version string reporting should be prettier.

Change-Id: Iff7fdde78bf2ba60a1ab88b38ca2ea89260eb5c7