Fix line length for pep8 compliance after cloning into other projects
There are a few lines that are close to the length limit so they pass
the PEP8 check under the common repository but fail the check after
the code is cloned into another project such as ceilometer. These
changes adjust the lines to give more space to allow the merge into
the other projects to go ahead without breaking the pep8 gating tests.

Change-Id: I33a05f5b51120114cc8ceffd0bf1b1ef462ed636
Signed-off-by: Doug Hellmann <doug.hellmann@dreamhost.com>