Make test_zmq bind/send to 127.0.0.1
By default, the impl_zmq driver binds to '*' (0.0.0.0);
For the purposes of testing, we only need to bind to 127.0.0.1.
Binding to all interfaces could be potentially dangerous, may simply not
work in some environments, etc. In particular, we do not want our
Jenkins slaves to be listening on public interfaces.

rpc_zmq_host is set to 127.0.0.1 as well, which avoids hostname
resolution. Normally, rpc_zmq_host is set to socket.hostname, which
normally should resolve to localhost, but we do not want to test that here.

Change-Id: I1224799bb399ceaf79088dd622be4b6823f29a41