Reduce test_zmq's reponse timeout.
The rpc response timeout in Nova's fake_flags.py was 5 seconds.
This was lost in the transition to openstack-common, is no longer
set globally. As in test_kombu.py, this should be set in test_zmq.py

The default value is much higher than it needs to be for the
purposes of testing. A lower value fails faster, resulting in faster
tests in the event of failures.

Change-Id: I106919d36acd19b702be85e453e949495db70ea9