Add healthcheck middleware from Swift
Swift contains a small piece of middleware that can be used as a first
port of call to determine if the service is responding.

To use it, one should prepend 'healthcheck' to the front of the
application pipeline, and add a filter section to the paste config that
looks something like the following:

[filter:healthcheck]
paste.filter_factory = foo.common.middleware.healthcheck:filter_factory

(where foo is the service in question, e.g. swift, glance, etc.)

Implements blueprint port-swift-healthcheck.

Change-Id: I2ccec53f879a760c10964f3b2cc99e5c1a6c8663