fanout subscriptions in impl_zmq
This necessary to support Quantum's scalable agents
in the Folsom release.

Supports matchmaking and subscribing to publishers,
in addition to sender-only fanout broadcasting.

Until now, senders of fanout messages had to know
of all recipients via the matchmaker. This patch
allows recipients to discover message publishers
via the matchmaker. This is useful in patterns
when there are few publishers, but many subscribers.

The raw message passed to ZeroMQ is changed
so that zmq_subscribe can filter by topic.

This patch may be partially reverted in Grizzly,
after the arrival of the ServiceGroupAPI which
will significantly improve the matchmaking
capabilities.

(cherry picked from commit 032bf345)

Change-Id: Ifc132519c6e380cb0a3643680764a38b841add0e