Don't trap then re-raise ImportError.
When we catch ImportError, print a message, and then raise a new ImportError,
we lose sane ability to read the traceback, and we gain nothing. Raising a new
ImportError on Attribute or Value error is helpful, given the context in which
we are working.

Change-Id: I52f61405e63535bcc745c9d350aa0dfbae36a8ac
(cherry picked from commit cd1424a5543ce01b3bd13887f2d1e818b8c296d3)