Code clean up
I fixed some PEP8 issues in nova/doc/ext/nova_todo.py and
nova/doc/source/conf.py in the past few days, intending to be
familiar with the bugfix workflow. But when I turned to the same
code files in other projects such as Swift, Keystone,... etc., I
found that they have the same problems, some even have issues
that beyond the scope of PEP8. So I decide to do a little code
clean up works on these files.
In this file, I tried to do things as follow:
1. PEP8 issues
2. Move the 'project' variable to the appropriate position.

Change-Id: Ifc2f8fdd2d3887e68b4d537d30d17b8c61cfba20