Update zmq context cleanup to use term.
Updates the RPC cleanup function for the zmq RPC driver
to use Context.term() instead of Context.destroy().

This allows the Nova code to work with older versions (pre 2.1.10)
of pyzmq. The Context term() function seems to do the job just as well as
destroy() (and is also thread safe) so lest just use it instead.

Fixes LP Bug #1044004.

Change-Id: Ifadd2c19914abfe5dbb58310d21eb6c78563fb82