Make projects define 'control_exchange'.
The 'control_exchange' option needs to have a project-specific default
value.  Just don't register this option and expect it to be registered
by the project using this code, at least for now.

   ****** IMPORTANT NOTE WHEN IMPORTING THIS CHANGE ******

If you are importing this change into a project that uses rpc, you must
add the control_exchange option in your code!

   *******************************************************

Change-Id: Ida5a8637c419e709bbf22fcad57b0f11c31bb959