Make projects define 'rpc_control_exchange'.
The first thing this patch does is to rename 'control_exchange' to
'rpc_control_exchange' to be more consistent with other rpc options.
The old name is still supported for backwards compatibility.

The 'rpc_control_exchange' option needs to have a project-specific default
value.  Just don't register this option and expect it to be registered
by the project using this code, at least for now.

   ****** IMPORTANT NOTE WHEN IMPORTING THIS CHANGE ******

If you are importing this change into a project that uses rpc, you must
add the rpc_control_exchange option in your code!  You should also
consider supporting 'control_exchange' as a deprecated alias for
backwards compatibility, as shown in the commented out example in this
patch.

   *******************************************************

Change-Id: Ida5a8637c419e709bbf22fcad57b0f11c31bb959