Change the qpid_heartbeat default to 60 seconds.
The heartbeat option enables sending a message at the AMQP layer at a
regular interval.  This helps ensure that the connection is still up and
working.  Once a failure is detected, a reconnection can be attempted.

The default for qpid_heartbeat was previously 5 seconds.  This is pretty
aggressive and can lead to invalid timeouts occuring.  An example is the
nova-compute service.  It may block in native code (such as libvirt)
under normal circumstances.  When this happens, other greenthreads (such
as the one responsible for sending the heartbeat) can not be scheduled.
This patch makes the default a less aggressive 1 minute.

Fix bug 1050661.

Change-Id: Idce48944050c331dda85e9b10ab6f5b025845c34
(cherry picked from commit 1a9d5c8df5aeee4faf99b54800566e386166836d)