Don't deprecate rabbit_host and rabbit_port
Where a single RabbitMQ broker node is used (rather than a cluster of
nodes) it is simplest for users to just set the rabbit_host option.
Often users will not need to specify a port number at all.

Similarly, where a cluster is used, it is simplest for the user to just
specify a list of node addresses in rabbit_hosts and assume that all
nodes are listening on the port specified in rabbit_port.

Change-Id: I4227541676dffc50cd6116225de0cec0afa36a41