Rewrite the policy engine from scratch
Complete rewrite of the policy engine, with careful thought given to
backwards-compatibility.  This rewrite, while supporting the old
list-of-lists syntax, has support for a new policy language having
"and", "or", and "not" operators.  In addition, support is available
for finer-grained policy decisions: policy checks can, using the
"case" construct, return values other than simply "True".  This
means that a policy rule can be written that allows for fine shades
of permissions; for instance, a rule could return the values,
"fulladmin", "projectadmin", or False, and the caller can use these
values to decide whether to allow all modifications, only some
modifications, or no modifications at all.

TODO:

 * Creation of a blueprint for feature tracking

Change-Id: I2951a0de3751bd2ec868e7a661070fed624e4af2