Rewrite the policy engine from scratch
Implements blueprint fine-grained-policy

Complete rewrite of the policy engine, with careful thought given to
backwards-compatibility.

Policy rules are now represented internally by a tree of Check
objects.

A new API is added for parsing and enforcing rules:

  set_rules(Rules.load_json(data, default_rule=...))
  result = check(rule, target, credentials)
  reset()

The old Brain/enforce API is deprecated and will be removed soon.

A new API is also added for registering new check types:

  @register("key")
  class KeyCheck(Check):
      def __call__(self, target, creds):
          return self.match in creds

Support for using functions as check types is deprecated and will
also be removed soon.

Change-Id: I2951a0de3751bd2ec868e7a661070fed624e4af2