Add a new policy language
Implements blueprint fine-grained-policy

Add a new policy language with "and" and "or" operators to replace
the old list-of-lists syntax. New '@' and '!' operators are also added.

This new language will enable us add more advanced features than the
old syntax would have allowed.

Backwards compat support for the old list-of-list syntax is retained.

Change-Id: I872cb6abf6f8051c3ff502a0fc7590cff4f63a25