Support RabbitMQ H/A cancellation notifications.
Upgrading to newer kombu and pyamqp transport to make
use of newer kombu's support for RabbitMQ cancellation
notifications, which happen when a consumer is consuming
from a replica of a queue and the queue's master dies.

With older kombu, in this case the consumer just never
gets any more messages even if the master recovers or if
messages are published to other replicas of the queue.

With new kombu, the consumer will get a ConsumerCancel
exception in connection.drain_events when this happens.

Newer kombu requires newer anyjson - see
https://bugs.launchpad.net/openstack-common/+bug/1031889

Alas, this change can only be tested manually so far.

Change-Id: I8242a33e0f9925feb0a6c3bf1d3f87832b8ceb8b