Revert "Add support for finer-grained policy decisions"
This reverts commit 3fc46892

After a productive discussion here:

  http://lists.openstack.org/pipermail/openstack-dev/2012-October/thread.html#1566

It appears that we don't yet have a really compelling use 
case for this and folks are worried about the extra 
complexity it brings. I think it's safe to say there's a
consensus that we shouldn't proceed with this idea yet. We 
can always revisit this again later if needs be.