Remove stale APIs.
A number of APIs were introduced into the openstack-common incubation
repo, but have since gone stable.  Nobody is working on them and we
don't want anyone trying to use them.  We should only include code here
that is actively being worked on and being moved toward an eventual
library release.  Otherwise, it should be removed.

Change-Id: I8eeb10fc9a21bc35b3004fa19782ebeeba39d20d