Fix a couple of file handle leaks, using with statements.
In practice unclosed file handle leaks are not a huge deal in CPython code,
because reference counting means they get closed automatically in a timely
manner.  But they're bad style, and they can become a real problem if the
code is run with a version of Python that doesn't use reference counting.

Change-Id: Ie54b979e26ffc9dd405871ee07c50b304cb36c67