Fix QPID reconnect issues
This fixes bug 1073999.

The problem results from the eventlet monkey patching and the way in
which qpid performs the session reconnection. The solution is to have
the qpid reconnection disbaled and the common roc code performing the
reconnection.

The time waiting between reconnections will be a multiple
of 2 starting at 1 and a maximum of 60 seconds.

Change-Id: I25067623481fcf29a6f949c3b884ea5830724d10