Fix QPID reconnect issues
The problem results from the eventlet monkey patching and the way in
which qpid performs the session reconnection. The solution is to have
the qpid reconnection disbaled and the common rpc code performing the
reconnection.

The time waiting between reconnections will be a multiple
of 2 starting at 1 and a maximum of 60 seconds.

This works around bug 1073999.

Change-Id: I3c53211d23e4e68ed73046cb3c637662ab73b339