Rename rabbit_notifier to rpc_notifier.
The previous rabbit_notifier module is generic and can be used
for all RPC notifications.

This commit moves the rabbit_notifier module to rpc_notifier and
adds a new deprecated rabbit_notifier module which can be used for Grizzly
so that users can have a chance to easily upgrade this config setting.

Once this lands here I'd like to do the same in Quantum.

Change-Id: I95073b1a8e06673d8e7363a4b37c050adbad97c5