Wrap eventlet's select because I don't trust it.
A bug was filed that showed services using impl_qpid locking up when
trying to connect to qpidd if qpidd was not running when the service was
started.  Digging into python-qpid showed that eventlet's select() was
saying that a pipe was ready for reading before anything had been
written to it, which resulted in doing a read() that blocked forever.

This patch patches the patched select().  It lets eventlet's select() do
its thing, but then does a non-blocking call to the system's built-in
select() to sanity check the result.  It should be harmless and come
with a neglible performance hit.  It does seem to work around the
observed issue.

Work around bug 1073999.

Change-Id: I89486aff8c50a247d5a54555af7b95a548ddf38f