Add module for loading specific classes
This adds loadables.py which contains code that is to be shared by
scheduling filtering and weighing in Nova and Cinder.  This is
cherry-picked from Nova (Change I1b217dc2bc2d1dc2235c8251318d06b46597e8f4)

Original commit message:

" ...
Most of this code originated from nova/scheduler/filters/__init__.py
(which was copied from nova/api extension loading).

This makes it more generic so that it can be shared. Note that this
functionality is quite different than the generic plugin manager that
exists in openstack-common. That code could not be used here without
some significant changes. It also seems pretty overkill for the
functionality required by scheduling filtering and weighing."

Change-Id: I7deeb0b6c78e4c81a41312eb9108ca1557ed71aa