Remove sitepackages=True from tox.ini
sitepackages=True violates the virtualenv as it causes system versions to
be used instead of testing what's in the pip-requires. We use it in nova
purely because we have no other way to get at the libvirt module. If we can
ever figure out another way around that, we will. It should certainly not
be used elsewhere.

Change-Id: Ieadc69d9872c78e1f53f1fdb17a09fe3f030cbf0