Add a CLI argument validation utility
blueprint oslo-cliutils

In nova-manage and cinder-manage, we allow command arguments to be
passed as optional or positional arguments.

e.g.

  $> nova-manage floating create 10.0.0.1/28
  $> nova-manage floating create --ip_range 10.0.0.1/28

are equivalent.

Once nova-manage has collected those arguments, it calls the
appropriate command function with them as positional and keyword
arguments. If the user forgets to supply a required argument, they
merely get a TypeError with little useful information.

Improve the usability of these commands using a new utility function
to check that the required arguments have been supplied and raise a
useful exception if not.

Change-Id: If6e4a9f222a30472bbfbcd06859865bd4e37e139