Provide i18n to those messages without _()
bug 1081498

There are a few log messages without _(), but a lot messages with _().
Wrap these log messages with _() in order to make the code consistent.

fix missing imports for _

Pass the log message and interpolation argument(s) to the logger separately

remove extra parens

rework the patch against latest trunk

Change-Id: Ib0a7b0f9c57aadfbc2cfc477c6123b885c9c6060