Use json instead of jsonutils in rpc.impl_fake.
The fake rpc backend uses json just to make sure the data that was
passed in can be serialized.  Our jsonutils module makes this check a
no-op, because it will automatically convert the data we pass in as
primitive types if needed.  Change this module back to using the json
module directly and add a comment about why we don't use jsonutils here.

Also add a test that ensures that non-primitive types raise an exception
in the fake rpc driver.

Change-Id: I61f46242cb98c875a94d86283e30efb24cc580ad