Fix broken --version command
The code for lazy loading of version string was broken. The way it
breaks in glance and oslo is different. This patch is just for oslo.
Here we end up overriding more methods in our lazy string class to make
argparse happy and added test cases to make sure we don't break again.
we also sure we cache the result of the lookup so we return the same
version everytime and for performance. Added a few test cases as well.

Fixes LP #1085984

Change-Id: I0635899c37fa8e5a7edf282907d2a5e04aa064cb