Fix qpid cast_to_server regression.
This patches addresses a regression in the qpid driver for the
cast_to_server and fanout_cast_to_server methods.  The regression was
introduced in 86952852c1dadd54f1294ae0adba01cfe827b8ce.

There are test cases that cover this code, but unfortunately they are
not being run by jenkins because python-qpid is not available in pypi.

Fix bug 1086848.

Change-Id: I89d4f81149f9a05c578b5ec544c5e34311484c6c