fix intermittent failures with deferred version test
Found a better mechanism to hook into the argparse's
for better control over the version printing mechanism
the problem was overriding string methods is hard and
flaky. it's better to use hooks in argparse to implement
the delayed lazy load of version information. Added a
test to try longer string and even run it in a loop.

Fixes LP #1087575

Change-Id: Ic510d67182d4f781856f2b1c1be0afbe9e28c7e1