fix intermittent failures with deferred version test
Need to get rid of the deferred string approach as
the problem was overriding string methods is hard and
flaky.  There is just no other documented
argparse hook to allow deferring of the version
calculation as version passed in has to be a str
or buffer. Added a test to try longer string and even
run it in a loop just to be sure.

Fixes LP #1087575

Change-Id: Icf83944d67b8ce5d3230627f3de10501da3276f2