Optionally add message envelope to notifications.
When message envelopes were added for rpc messages, notifications were
explicitly left out.  This was due to the fact that notifications are
consumed outside of OpenStack.  However, notification consumers will
likely want to eventually start taking advantage of some of the features
that the envelope is a prerequisite for, such as message signing.

This patch adds YACO (yet another configuration option) that allows
administrators to enable the message envelope for notifications.  It is
off by default (for now, at least).

Related to blueprint version-rpc-messages.

Change-Id: Iaa8b437e6b4f64053f4b02bf486516f60bb020e9