Move rootwrap code to openstack.common
Copies current nova-rootwrap code to openstack.common, so that it can be
reused by Cinder and Quantum. Before it can be used in projects,
update.py needs to grow the capability to deploy files in bin/ and etc/,
as well as replacing a placeholder text by the destination project name
in source files and binary names.

In this proposed version, the placeholder text is "oslo".

Change-Id: I8655d5b3cccacd1cc2225aa539339fb478615422