This patch allows the --version switch to work properly for CLI programs.
The _deferred_version_string is passed into optparse.OptionParser as
a version string where it is treated as a string.  optparse.OptionParser
calls replace() on the string.  However, because it is not overridden
in _deferred_version_string the wrong string is created and the version
information is not retained.  This patch fixes that problem

Unlike bug #1092631 and exception is not thrown in this case, however
inaccurate version information is displayed.

Change-Id: I4c571ce22d6f3e72e4077cf18698fed645629c17