Ensure proxy callback threads exit during tests
On the consumer side, each message is consumed in a separate thread
but we currently have no way of waiting for all these threads to
complete.

In BaseRpcTestCase.test_call_timeout(), we complete the thread while
there is still a message handler running on the other side. This can
mean the consumer attempting to send its reply while other tests
are running, which clearly is a bad thing.

Make the connection close() and reset() methods in the kombu and qpid
drivers wait on proxy object threads before returning to avoid this
situation.

Change-Id: I8fb84b1dc7f31196ef27a5d1f8c4a3346b03e0ae