Fix serialization in impl_zmq.
When the rpc message envelope patch went in, the serialization done in
impl_zmq was replaced with what is done in rpc.common.  Instead, it just
needs to be done as an additional stage of serialization.  The reason is
that the amqp drivers can handle passing dicts into the messaging
libraries.  The zeromq library can not handle that, so it has to ensure
it is a string before passing it down to the library.

This patch restores the original (unmodified) _serialize() and
_deserialize() methods in impl_zmq that were removed and calls them as
appropriate.

Change-Id: I09cac780c9a9b15390229c63d71e360120548b3d