Fix timezone handling in timeutils tests
The tests for timeutils use some canned datetime values that include
timezones, and override the result of utcnow() with the same. This
is not correct because the real utcnow() returns naive datetime
objects, so tests that assume the contrary are incorrect.

This came about because I was unable to satisfy both the oslo unit
tests and the nova ones with the same code. That happened because
the oslo ones override utcnow() with an invalid result, and the
nova tests use the real one.

Change-Id: I683818f77ce0d3e3b2f07c0326d7a7196f3346c3