Add testr and testr_coverage comands.
We have a couple of chunks of shell that are repeated in all of the
tox.ini files for the projects that are using testr. Not only is it repeated,
but the use of shell in the tox.ini file causes tox itself to print a warning
which is distracting. Encapsulate that into setup.py commands so that:
 a) it's run via python in the venv, silencing tox warning
 b) we don't have to copy the code all over the place

Change-Id: I637156d26a1d63e960939a2818f45ea88042e9b6