Fix zmq socket.close() with eventlet 0.9.17
A bug/change in eventlet 0.9.17 blocks the
kwargs from close() from passing through
to pyzmq. As this code was only explicitly
passing defaults, no argument is necessary.

This addresses file descriptor leaks reported
in bug 1099185. This bug does not affect
installations with eventlet<=0.9.16

The code will now log an error if a socket
cannot be closed.

Change-Id: Id438dab9744042328b7fac44693d02e116d8c1d9