blueprint amqp-rpc-fast-reply-queue AMQP RPC single per process response queue performance improvement For AMQP based RPC, specifically RabbitMQ and Qpid, this change replace the dynamically created RPC call reply queue with a single queue that is created on teh first RPC call and used on all subsequent calls. It provides backward compatibility on the callee side by recognizing downlevel callers and on the caller side by addint a config option to revert to the old dynamically created queue based upon the msg_id.
Change-Id: Idb09a71472866bd3950f58d4f7f45a3181eb40fc

blueprint amqp-rpc-fast-reply-queue

For AMQP based RPC, specifically RabbitMQ and Qpid, this change replace
the dynamically created RPC call reply queue with a single queue that
is created on teh first RPC call and used on all subsequent calls.
It provides backward compatibility on the callee side by recognizing
downlevel callers and on the caller side by addint a config option to
revert to the old dynamically created queue based upon the msg_id.

Change-Id: Idb09a71472866bd3950f58d4f7f45a3181eb40fc

blueprint amqp-rpc-fast-reply-queue

For AMQP based RPC, specifically RabbitMQ and Qpid, this change replace
the dynamically created RPC call reply queue with a single queue that
is created on teh first RPC call and used on all subsequent calls.
It provides backward compatibility on the callee side by recognizing
downlevel callers and on the caller side by addint a config option to
revert to the old dynamically created queue based upon the msg_id.

Change-Id: Idb09a71472866bd3950f58d4f7f45a3181eb40fc