Don't catch GreenletExit with "except Exception:"
GreenletExit is, both by name and definition, an exit exception. The
Exception class is the "Common base class for all non-exit exceptions",
however GreenletExit is implemented as a subclass of Exception.  We do
not wish to catch exit exceptions in "except Exception:" clauses, so
monkey-patch GreenletExit to inherit directly from BaseException as it
should.

bug 1097203

Change-Id: Ia91efcbe69d9162140a9fb2994ec947f71b2f652
Signed-off-by: Zane Bitter <zbitter@redhat.com>