Leverage style & eliminate msg unpacking.
The style field has been used and reserved.
This patch passes the message type
over the style field and uses this to
optimize message processing.

The message unpacking on the proxy
has proven to be unnecessary. This is
a fairly slow operation, so we should
avoid doing it.

Additionally, packing and unpacking
the message has security implications
once we move to trusted messaging.

Also removes the passing of msg_id by
non-call/reply messages as it was
redundant and not necessary for
backwards compatiblity.

Change-Id: Idbc9291fcce6ff6f6108573811d065a5c2c24255

Fixes to msg unpacking and style

Change-Id: I9d6a1483afaed59462e208102448acccc02cff63