Fix Lack of reconnecting operation when using rabbitmq mirrored queue
On using Mirrored Queue feature in impl_kombu.py, there is a case
that messages are lost because amqp client does not handle exceptions
properly when rabbitmq is dead on the way to wait return value of call().
This patch fix this and enables ampq client reconnecting slave rabbitmq.

Fixes bug 1102051

Change-Id: Ia7a1b9067f7ea4639195a1548de29e0364368e51